backend/requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
Markdown==3.4.3
py4j==0.10.9.7
Pygments==2.15.1
pyspark==3.4.1
pytz==2023.3
sqlparse==0.4.4
